Make the library compatible with Symfony >= 2.3
